+++ breeze-core/src/main/kotlin/Named.kt
@@ -1,0 +1,11 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.core
+
+/**
+ * 拥有名字的对象。
+ */
+interface Named<T:CharSequence> {
+	val name:T
+}