+++ breeze-core/src/main/kotlin/Validated.kt
@@ -1,0 +1,13 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.core
+
+/**
+ * 可验证是否合法的对象。
+ */
+interface Validated {
+	fun validate(): Boolean {
+		return true
+	}
+}